updates : - making core module to set common sources - update .gitignore to ignore entire .idea - add packages to tidy up files - add onClick method to ArtikelFragment.kt by inherit from OnClickListener.kt
