+++ .idea/misc.xml
@@ -4,27 +4,18 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/drawable/bg_header.xml" value="0.2985" />
-        <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/drawable/correct_question_option.xml" value="0.2985" />
-        <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/drawable/desain_image.xml" value="0.196" />
-        <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/drawable/question_option.xml" value="0.2985" />
-        <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/drawable/selected_question_option.xml" value="0.2985" />
-        <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/drawable/wrong_question_option.xml" value="0.2985" />
-        <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/layout/activity_artikel.xml" value="0.17831813576494426" />
         <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/layout/activity_home.xml" value="0.18802083333333333" />
         <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/layout/activity_home2.xml" value="0.18802083333333333" />
         <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/layout/activity_login.xml" value="0.22" />
         <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/layout/activity_main.xml" value="0.36666666666666664" />
-        <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/layout/activity_pre_test.xml" value="0.25" />
         <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/layout/activity_register.xml" value="0.18802083333333333" />
-        <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/layout/activity_result.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/layout/activity_sign_up.xml" value="0.21" />
         <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/layout/activity_splash_screen.xml" value="0.2684563758389262" />
-        <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/layout/fragment_artikel.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/layout/fragment_home.xml" value="0.18802083333333333" />
         <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/layout/fragment_profile.xml" value="0.17395833333333333" />
-        <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/layout/list_item.xml" value="0.18802083333333333" />
         <entry key="..\:/Users/sony/AndroidStudioProjects/SelfCare/app/src/main/res/menu/bottom_nav_menu.xml" value="0.18802083333333333" />
         <entry key="app/src/main/res/layout/activity_home2.xml" value="0.35625" />
+        <entry key="app/src/main/res/layout/activity_pre_test.xml" value="0.35625" />
       </map>
     </option>
   </component>
