--- app/src/main/java/com/example/self_care/QuestionData.kt
@@ -1,12 +1,0 @@
-package com.example.self_care
-
-data class QuestionData (
-    var id:Int,
-    var question:String,
-
-    var option_one:String,
-    var option_tw0:String,
-    var option_three:String,
-    var option_four:String,
-    var correct_ans:Int
-)
