+++ app/src/main/java/com/example/self_care/domain/Artikel.kt
@@ -1,3 +1,3 @@
-package com.example.self_care
+package com.example.self_care.domain
 
 data class Artikel(var titleImage: Int, var heading:String)
