+++ app/src/main/java/com/example/self_care/view/artikel/ArtikelActivity.kt
@@ -1,11 +1,12 @@
-package com.example.self_care
+package com.example.self_care.view.artikel
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.self_care.R
 
-class HomeActivity : AppCompatActivity() {
+class ArtikelActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_home)
+        setContentView(R.layout.activity_artikel)
     }
 }
