+++ app/src/main/java/com/example/self_care/view/auth/LoginActivity.kt
@@ -1,9 +1,11 @@
-package com.example.self_care
+package com.example.self_care.view.auth
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.TextView
+import com.example.self_care.R
+import com.example.self_care.view.home.HomeActivity2
 
 class LoginActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
