+++ app/src/main/java/com/example/self_care/view/auth/SignUpActivity.kt
@@ -1,9 +1,10 @@
-package com.example.self_care
+package com.example.self_care.view.auth
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.TextView
+import com.example.self_care.R
 
 class SignUpActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
