+++ app/src/main/java/com/example/self_care/view/auth/SplashScreen.kt
@@ -1,9 +1,10 @@
-package com.example.self_care
+package com.example.self_care.view.auth
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
+import com.example.self_care.R
 
 class SplashScreen : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
