+++ app/src/main/java/com/example/self_care/view/home/HomeActivity.kt
@@ -1,11 +1,12 @@
-package com.example.self_care
+package com.example.self_care.view.home
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.self_care.R
 
-class ArtikelActivity : AppCompatActivity() {
+class HomeActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_artikel)
+        setContentView(R.layout.activity_home)
     }
 }
