+++ app/src/main/java/com/example/self_care/view/home/HomeActivity2.kt
@@ -1,9 +1,12 @@
-package com.example.self_care
+package com.example.self_care.view.home
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.fragment.app.Fragment
+import com.example.self_care.ProfileFragment
+import com.example.self_care.R
 import com.example.self_care.databinding.ActivityHome2Binding
+import com.example.self_care.view.artikel.ArtikelFragment
 
 class HomeActivity2 : AppCompatActivity() {
