+++ app/src/main/java/com/example/self_care/view/profile/ProfileFragment.kt
@@ -1,12 +1,11 @@
-package com.example.self_care
+package com.example.self_care.view.profile
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
+import com.example.self_care.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -15,21 +14,13 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [ArtikelFragment.newInstance] factory method to
+ * Use the [ProfileFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class ArtikelFragment : Fragment() {
+class ProfileFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
-
-    private lateinit var adapter : AdapterArtikel
-    private lateinit var recyclerView: RecyclerView
-    private lateinit var artikelArrayList : ArrayList<Artikel>
-
-    lateinit var imageId : Array<Int>
-    lateinit var heading : Array<String>
-    lateinit var artikel : Array<String>
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -44,21 +35,7 @@ class ArtikelFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-
-        return inflater.inflate(R.layout.fragment_artikel, container, false)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        dataInitialize()
-        val layoutManager = LinearLayoutManager(context)
-        recyclerView = view.findViewById(R.id.recyclerView)
-        recyclerView.layoutManager = layoutManager
-        recyclerView.setHasFixedSize(true)
-        adapter = AdapterArtikel(artikelArrayList)
-        recyclerView.adapter = adapter
-
-
+        return inflater.inflate(R.layout.fragment_profile, container, false)
     }
 
     companion object {
@@ -68,28 +45,16 @@ class ArtikelFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment ArtikelFragment.
+         * @return A new instance of fragment ProfileFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            ArtikelFragment().apply {
+            ProfileFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
                 }
             }
-    }
-
-    private fun dataInitialize(){
-        artikelArrayList = arrayListOf<Artikel>()
-
-
-
-        for (i in imageId.indices){
-            val artikel = Artikel(imageId[i],heading[i])
-            artikelArrayList.add(artikel)
-        }
-
     }
 }
