+++ app/src/main/res/layout/activity_login.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="#EAEBEE"
     app:layout_constraintVertical_bias="0"
-    tools:context=".LoginActivity">
+    tools:context=".view.auth.LoginActivity">
 
     <androidx.appcompat.widget.LinearLayoutCompat
         android:layout_width="match_parent"
