+++ app/src/main/res/layout/activity_result.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:layout_gravity="center"
     android:background="#EAEBEE"
-    tools:context=".ResultActivity">
+    tools:context=".view.test.ResultActivity">
 
     <androidx.appcompat.widget.LinearLayoutCompat
         android:layout_width="match_parent"
