+++ app/src/main/res/layout/activity_sign_up.xml
@@ -6,7 +6,7 @@
 
     android:background="#EAEBEE"
     android:layout_height="match_parent"
-    tools:context=".SignUpActivity">
+    tools:context=".view.auth.SignUpActivity">
 
     <androidx.appcompat.widget.LinearLayoutCompat
         android:layout_width="match_parent"
