+++ app/src/main/res/layout/activity_splash_screen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:background="#EAEBEE"
     android:layout_height="match_parent"
-    tools:context=".SplashScreen">
+    tools:context=".view.auth.SplashScreen">
 
     <androidx.appcompat.widget.LinearLayoutCompat
         android:layout_width="match_parent"
