+++ app/src/main/res/layout/fragment_artikel.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ArtikelFragment">
+    tools:context=".view.artikel.ArtikelFragment">
 
 <!--    &lt;!&ndash; TODO: Update blank fragment layout &ndash;&gt;-->
 <!--    <TextView-->
