+++ core/src/main/java/com/github/core/listener/OnClickListener.kt
@@ -1,0 +1,5 @@
+package com.github.core.listener
+
+interface OnClickListener {
+    fun onClick(data: Any, position: Int)
+}
